Fonts dir checks if already exists

+++ run_once_before_install-packages.sh
@@ -64,12 +64,15 @@ flatpak install -y $(tr '\n' ' ' < "$pkgmgmt/flatpak.list")
 flatpak uninstall --unused
 
 # Install Jetbrains Mono Nerd Font
-echo "Installing JetBrains Mono Nerd Font..."
-mkdir -p "$HOME/.local/share/fonts"
-cd "$HOME/.local/share/fonts"
-wget -q --show-progress "https://github.com/ryanoasis/nerd-fonts/releases/latest/download/JetBrainsMono.zip"
-unzip -o JetBrainsMono.zip
-rm JetBrainsMono.zip
+fonts_dir="$HOME/.local/share/fonts"
+if [ ! -d "$fonts_dir/JetBrains" ]; then
+    echo "Installing JetBrains Mono Nerd Font..."
+    mkdir -p "$fonts_dir"
+    cd "$fonts_dir"
+    wget -q --show-progress "https://github.com/ryanoasis/nerd-fonts/releases/latest/download/JetBrainsMono.zip"
+    unzip -o JetBrainsMono.zip
+    rm JetBrainsMono.zip
+fi
 
 # Refresh font cache
 fc-cache -fv
@@ -79,7 +82,7 @@ echo "JetBrains Mono Nerd Font installed successfully!"
 curl -s https://ohmyposh.dev/install.sh | bash -s
 
 # eCalc
-mkdir "$HOME/work/github"
+mkdir -p "$HOME/work/github"
 cd "$HOME/work/github"
 git clone git@github.com:bekesiv/ecalc.git
 cd install
